+++ CHANGES.md
@@ -3,8 +3,7 @@
 - Fixed a paragraph formatting issue with large inline images (Issue #369)
 - Fixed a buffer underflow issue (Issue #370)
 - Added support for a new `L` header/footer format (`$LETTERHEAD`), which
-  inserts the logo image at its full size to be used as a letterhead
-  (Issue #372)
+  inserts a letterhead image at its full size (Issue #372, Issue #373)
 - Updated the build documentation (Issue #374)