+++ CHANGES.md
@@ -6,7 +6,8 @@
 - Fixed a buffer underflow issue (Issue #370)
 - Fixed PDF page numbers (Issue #371)
 - Added support for a new `L` header/footer format (`$LETTERHEAD`), which
-  inserts a letterhead image at its full size (Issue #372, Issue #373)
+  inserts a letterhead image at its full size (Issue #372, Issue #373,
+  Issue #375)
 - Updated the build documentation (Issue #374)